Move DisplayNames to core too.
