+++ src/libraries/Microsoft.PowerFx.Core/Public/Engine.cs
@@ -6,6 +6,8 @@ using System.Linq;
 using Microsoft.PowerFx.Core;
 using Microsoft.PowerFx.Core.Binding;
 using Microsoft.PowerFx.Core.Glue;
+using Microsoft.PowerFx.Core.Lexer;
+using Microsoft.PowerFx.Core.Localization;
 using Microsoft.PowerFx.Core.Parser;
 using Microsoft.PowerFx.Core.Public;
 using Microsoft.PowerFx.Core.Public.Types;
@@ -147,6 +149,58 @@ namespace Microsoft.PowerFx
             var suggestions = intellisense.Suggest(context, binding, formula);
 
             return suggestions;
+        }
+
+        /// <summary>
+        /// Convert references in an expression to the invariant form.
+        /// </summary>
+        /// <param name="expressionText">textual representation of the formula.</param>
+        /// <param name="parameters">Type of parameters for formula. The fields in the parameter record can 
+        /// be acecssed as top-level identifiers in the formula. If DisplayNames are used, make sure to have that mapping
+        /// as part of the RecordType.
+        /// <returns>The formula, with all identifiers converted to invariant form</returns>
+        public string GetInvariantExpression(string expressionText, RecordType parameters)
+        {
+            return ConvertExpression(expressionText, parameters, toDisplayNames: false);
+        }
+
+        /// <summary>
+        /// Convert references in an expression to the display form.
+        /// </summary>
+        /// <param name="expressionText">textual representation of the formula.</param>
+        /// <param name="parameters">Type of parameters for formula. The fields in the parameter record can 
+        /// be acecssed as top-level identifiers in the formula. If DisplayNames are used, make sure to have that mapping
+        /// as part of the RecordType.
+        /// <returns>The formula, with all identifiers converted to display form</returns>
+        public string GetDisplayExpression(string expressionText, RecordType parameters)
+        {
+            return ConvertExpression(expressionText, parameters, toDisplayNames: true);
+        }
+
+        private string ConvertExpression(string expressionText, RecordType parameters, bool toDisplayNames)
+        {
+            var formula = new Formula(expressionText);
+            formula.EnsureParsed(TexlParser.Flags.None);
+
+            var resolver = CreateResolver();
+            var binding = TexlBinding.Run(
+                new Glue2DocumentBinderGlue(),
+                null,
+                new Core.Entities.QueryOptions.DataSourceToQueryOptionsMap(),
+                formula.ParseTree,
+                resolver,
+                ruleScope: parameters._type,
+                useThisRecordForRuleScope: false,
+                updateDisplayNames: toDisplayNames,
+                forceUpdateDisplayNames: toDisplayNames);
+
+            Dictionary<Span, string> worklist = new ();
+            foreach (var token in binding.NodesToReplace)
+            {
+                worklist.Add(token.Key.Span, TexlLexer.EscapeName(token.Value));
+            }
+
+            return Span.ReplaceSpans(expressionText, worklist);
         }
     }
 }
