+++ src/libraries/Microsoft.PowerFx.Interpreter/RecalcEngine.cs
@@ -158,58 +158,6 @@ namespace Microsoft.PowerFx
             return newValue;
         }
 
-        /// <summary>
-        /// Convert references in an expression to the invariant form.
-        /// </summary>
-        /// <param name="expressionText">textual representation of the formula.</param>
-        /// <param name="parameters">Type of parameters for formula. The fields in the parameter record can 
-        /// be acecssed as top-level identifiers in the formula. If DisplayNames are used, make sure to have that mapping
-        /// as part of the RecordType.
-        /// <returns>The formula, with all identifiers converted to invariant form</returns>
-        public string GetInvariantExpression(string expressionText, RecordType parameters)
-        {
-            return ConvertExpression(expressionText, parameters, toDisplayNames: false);
-        }
-
-        /// <summary>
-        /// Convert references in an expression to the display form.
-        /// </summary>
-        /// <param name="expressionText">textual representation of the formula.</param>
-        /// <param name="parameters">Type of parameters for formula. The fields in the parameter record can 
-        /// be acecssed as top-level identifiers in the formula. If DisplayNames are used, make sure to have that mapping
-        /// as part of the RecordType.
-        /// <returns>The formula, with all identifiers converted to display form</returns>
-        public string GetDisplayExpression(string expressionText, RecordType parameters)
-        {
-            return ConvertExpression(expressionText, parameters, toDisplayNames: true);
-        }
-
-        private string ConvertExpression(string expressionText, RecordType parameters, bool toDisplayNames)
-        {
-            var formula = new Formula(expressionText);
-            formula.EnsureParsed(TexlParser.Flags.None);
-
-            var resolver = new RecalcEngineResolver(this, Config);
-            var binding = TexlBinding.Run(
-                new Glue2DocumentBinderGlue(),
-                null,
-                new Core.Entities.QueryOptions.DataSourceToQueryOptionsMap(),
-                formula.ParseTree,
-                resolver,
-                ruleScope: parameters._type,
-                useThisRecordForRuleScope: false,
-                updateDisplayNames: toDisplayNames,
-                forceUpdateDisplayNames: toDisplayNames);
-
-            Dictionary<Span, string> worklist = new ();
-            foreach (var token in binding.NodesToReplace)
-            {
-                worklist.Add(token.Key.Span, TexlLexer.EscapeName(token.Value));
-            }
-
-            return Span.ReplaceSpans(expressionText, worklist);
-        }
-
         // Invoke onUpdate() each time this formula is changed, passing in the new value. 
         public void SetFormula(string name, string expr, Action<string, FormulaValue> onUpdate)
         {
